requirements.txt
Django==4.2.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
openai==0.27.8
django-crispy-forms==2.0
crispy-bootstrap4==2022.1
pillow==9.5.0
python-frontmatter==1.0.0
markdown==3.4.3
reportlab==3.6.12
xhtml2pdf==0.2.8